move JoinCommon funcs out of IJoin.h/cpp
